Edit stdoutService
