+++ src/handlers/stdoutLineHandler.js
